Moved the structs to another file, and verified that it works
